Refactored out old graphql.ts file
